[스프링 - 협업 미션] 소롱(최솔지) 미션 제출합니다. (#23)
* feat: application 설정

* fix: 테이블명 수정

* fix: application 설정 변경

* fix: application 설정 변경

* fix: applcation 설정 변경, test 추가

* fix: schema.sql 수정

* feat: rest uri에 api 경로 추가

* feat: 문서화 관련 의존성 추가

* feat: 문서화 관련 설정

- AuthenticationPrincipalConfig 를 test에서 사용하지 않게 변경
- 문서화 설정 추가

* feat: 역 생성, 조회 테스트와 문서화

* feat: 역 삭제 테스트와 문서화

* feat: 역 관리, 경로 관리 테스트와 문서화

- 테스트 데이터를 위한 TestDataLoader 생성
- LineController에서 delete 메서드를 noContent 반환하도록 변경
- 노선 수정을 위해 LineUpdateRequest 추가

* refactor: AuthenticationPrincipalConfig 수정

* refactor: 빈 생성자 추가

* refactor: application-test.yml 테스트 패키지로 이동

* refactor: 노선 삭제 바뀐 응답값 적용

* feat: 인증 테스트 작성과 문서화

* refactor: 유저 수정 시 유저 정보를 반환하도록 변경

* feat: 유저 관련 테스트와 문서화

* refactor: DataLoader 수정

- 테이블이 비어있을 때만 데이터 삽입하도록 변경
- schema.sql 에서 drop 쿼리 제거

* refactor: api 문서 가독성 있게 수정

* refactor: index.adoc 수정

* refactor: 노선 조회 시 구간 정보를 주도록 변경

- SectionResponse 생성
- Section을 정렬하는 로직 추가
- LineResponse가 SectionResponse를 갖도록 수정
- 관련된 테스트 수정
- API 문서 수정

* feat: 이메일 중복 확인 API 추가

- 중복된 이메일이 있을 때 false 반환
- findByEmail 실패했을 때 예외 처리
- API 문서 수정

* fix: 등록된 구간이 없을때 NPE 해결

* refactor: API 문서 출력 개선

* feat: 역 수정 API, 역 이름 중복 예외 처리 추가

- 테스트 작성 및 수정
- API 문서 개선

* refactor: API 문서 수정 및 출력 개선

* refactor: 노선 수정 시 노선 객체 응답하도록 변경

* refactor: InvalidPathException 분리

* feat: Authorization Interceptor 적용

- 인터셉터에서 사용자 인증 관리
- 조회 외의 기능은 모두 인증 필요
- Authority enum 추가
- WebConfig에 인증 요청 할 uri 추가

* feat: Authorization Interceptor 적용에 따른 테스트 수정

- 테스트용 유저 테스트 DB에 삽입
- 인증 통과하는 로직 추가
- 인수 테스트에서 토큰 요청하는 로직 추가

* refactor: transactional 설정

* feat: 인증 실패 테스트 추가

* fix: 테스트 수정

* refactor: transactional 설정

* Interceptor 관련 설정 변경
- Options 메서드 승인
- api 요청 uri 추가

* fix: 실패하는 테스트 수정
- 인증 통과하도록 변경

* refactor: 노선을 삭제하면 구간까지 삭제되도록 수정

* feat: 구간에 등록된 역은 삭제 불가하도록 변경

* refactor: 노선, 역 검증 추가
- 중복된 이름으로 생성 불가
- 예외 처리 추가

* refactor: 역 추가 시 이름 중복 검증

* refactor: transactional 수정

* feat: Request 값 유효성 검증 추가

- Valid annotation 만들어서 사용
- Exceptionhandler 추가

* feat: 경로 조회 시 요금 정보 기능 구현

* refactor: Valid annotation 수정

- 실패하는 테스트 수정

* feat: 노선별 추가 요금 기능 구현

* feat: 연령별 요금 계산 기능 구현

- 테스트 추가 및 수정

* refactor: 역 목록 조회 시 속한 노선 정보 함께 반환

* refactor: 노선 조회 시 역과 거리 정보를 함께 반환

* feat: Station 변경 내용 API 문서 적용

* style: 코드 컨벤션 적용

* refactor: Dao 개선
- 배열 생성 제거
- 언박싱 제거

* refactor: rowMapper 메서드명 변경 및 람다식으로 통일

* refactor: @Transactional 수정
- DB 변경이 일어나는 경우에만 사용하도록 변경

* refactor: 도메인 생성자 변경
- 주 생성자를 하나만 두고 부 생성자 작성

* refactor: Pattern 인스턴스 캐싱하도록 분리

* refactor: StationWithDistanceResponse의 기본값 변경
- null 일때 표시하지 않도록 변경

* refactor: RuntimeException 수정
- 상황에 맞는 예외 만들어서 처리

* refactor: 제네릭 리스트 반환하도록 변경

* refactor: Custom Exception 추상화
- HttpStatus를 exception에서 처리
- ExceptionHandler 간략화

* refactor: 개행과 생성자 수정

* refactor: 불필요한 코드제거

* refactor: HttpStatus 변경

* refactor: 테스트시 회원 등록 위치 변경
- 토큰 요청할 때가 아닌 테스트마다 하도록 수정

* fix: 지하철 역 관련 테스트 실패 시 응답 수정

* fix: JSON 변환시 NPE 해결

* fix: 예외 처리 수정

* refactor: countSectionByStationId 반환값 수정

* refactor: SubwayException 제한자 추가

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,23 +1,41 @@
 package wooteco.subway.station.application;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.exception.badrequest.CannotRemoveStationException;
+import wooteco.subway.exception.duplicate.DuplicatedStationNameException;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.dao.SectionDao;
+import wooteco.subway.line.domain.Lines;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
+import wooteco.subway.station.dto.SimpleLineResponse;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
+@Transactional
 public class StationService {
-    private StationDao stationDao;
 
-    public StationService(StationDao stationDao) {
+    private static final int ZERO_COUNT = 0;
+    private StationDao stationDao;
+    private SectionDao sectionDao;
+    private LineDao lineDao;
+
+    public StationService(StationDao stationDao, SectionDao sectionDao,
+        LineDao lineDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
+        this.lineDao = lineDao;
     }
 
+    @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
+        String name = stationRequest.getName();
+        validateDuplicate(name);
+
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
@@ -26,15 +44,39 @@ public class StationService {
         return stationDao.findById(id);
     }
 
-    public List<StationResponse> findAllStationResponses() {
+    public List<StationResponse> findAllStations() {
         List<Station> stations = stationDao.findAll();
+        Lines lines = new Lines(lineDao.findAll());
 
         return stations.stream()
-                .map(StationResponse::of)
-                .collect(Collectors.toList());
+            .map(station -> StationResponse.of(
+                station,
+                SimpleLineResponse.listOf(lines.getLineContainStation(station))
+            ))
+            .collect(Collectors.toList());
     }
 
+    @Transactional
     public void deleteStationById(Long id) {
+        if (sectionDao.countSectionByStationId(id) > ZERO_COUNT) {
+            throw new CannotRemoveStationException();
+        }
         stationDao.deleteById(id);
+    }
+
+    @Transactional
+    public StationResponse updateStation(Long id, StationRequest stationRequest) {
+        String name = stationRequest.getName();
+        validateDuplicate(name);
+
+        Station station = new Station(id, name);
+        stationDao.update(station);
+        return StationResponse.of(station);
+    }
+
+    private void validateDuplicate(String name) {
+        if (stationDao.findByName(name).isPresent()) {
+            throw new DuplicatedStationNameException();
+        }
     }
 }
